backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
duckduckgo-search
apscheduler
pydantic
python-dotenv
langchain-communitypraw
newsapi-python
tavily-python
